requirements.txt
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.0
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
pycparser==2.22
PySocks==1.7.1
selenium==4.23.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
gunicorn==22.0.0
Flask==2.2.2
Werkzeug==2.2.2
qrcode==7.4.2
python-barcode==0.15.1
Pillow==9.4.0
matplotlib==3.7.0
numpy==1.23.5
